explicit cast

+++ code/particle/particle.cpp
@@ -136,7 +136,7 @@ namespace particle
 		}
 
 		// treat particles on lower detail levels as 'further away' for the purposes of culling
-		float adjusted_dist = vm_vec_dist(&Eye_position, &info->pos) * powf(2.5f, (NUM_DEFAULT_DETAIL_LEVELS - Detail.num_particles));
+		float adjusted_dist = vm_vec_dist(&Eye_position, &info->pos) * powf(2.5f, (float)(NUM_DEFAULT_DETAIL_LEVELS - Detail.num_particles));
 		// treat bigger particles as 'closer'
 		adjusted_dist /= info->rad;
 		float cull_start_dist = 1000.f;
